@@ -1,35 +1,29 @@
 {
-  "name": "pixelpokerserver",
-  "version": "0.0.1",
-  "description": "Server for Pixel poker game",
-  "main": "app.ts",
-  "author": "wmbryce",
-  "license": "MIT",
-  "dependencies": {
-    "aes-js": "^3.1.2",
-    "colors": "^1.4.0",
-    "cors": "^2.8.5",
-    "express": "^4.17.2",
-    "lodash": "^4.17.21",
-    "socket.io": "^4.4.1",
-    "ts-node": "^10.4.0",
-    "typescript": "^4.5.4",
-    "uuid": "^8.3.2"
-  },
+  "name": "server",
+  "version": "1.0.0",
+  "private": true,
   "scripts": {
-    "prebuild": "eslint eslintrc.json tsconfig.json --fix",
-    "build": "tsc",
-    "prestart": "npm run build",
-    "start": "nodemon app.ts",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "dev": "bun --watch app.ts",
+    "build": "bun build app.ts --outdir dist --target bun",
+    "start": "bun app.ts",
+    "test": "bun test"
+  },
+  "dependencies": {
+    "@pixelpoker/shared": "workspace:*",
+    "chalk": "^5.3.0",
+    "cors": "^2.8.5",
+    "express": "^4.19.2",
+    "lodash": "^4.17.21",
+    "pokersolver": "^2.1.4",
+    "socket.io": "^4.7.5",
+    "uuid": "^10.0.0"
   },
   "devDependencies": {
-    "@tsconfig/node12": "^1.0.9",
-    "@types/express": "4",
-    "@types/node": "16",
-    "@typescript-eslint/eslint-plugin": "^5.10.0",
-    "@typescript-eslint/parser": "^5.10.0",
-    "eslint": "^8.7.0",
-    "nodemon": "^2.0.15"
+    "@types/cors": "^2.8.17",
+    "@types/express": "^4.17.21",
+    "@types/lodash": "^4.17.7",
+    "@types/node": "^22.0.0",
+    "@types/uuid": "^10.0.0",
+    "typescript": "^5.5.4"
   }
 }
