@@ -1,139 +1,81 @@
-import React, { useState, useEffect, useRef } from "react";
-import styled from "@emotion/styled";
+import { useState, useEffect, useRef } from 'react';
+import socket from '../socket';
+import type { ChatMessage } from '@pixelpoker/shared';
 
-//gets the data from the action object and reducers defined earlier
 interface Props {
-  username?: String;
-  room?: String;
-  Socket: any;
+  username: string;
+  room: string;
 }
-const ChatBoxContainer = styled.div`
-  width: 90%;
-  height: 100%;
-  margin: 10px 10px 10px 10px;
-  padding: 10px 10px 10px 10px;
-`;
 
-const MessageRowRecieved = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: flex-start;
-  border: 1px solid grey;
-  border-radius: 5px;
-  margin: 10px 300px 10px 0px;
-  padding: 0px 20px;
-  //width: 80%;
-`;
-
-const MessageRowSent = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-end;
-  align-items: flex-end;
-  border: 1px solid grey;
-  border-radius: 5px;
-  margin: 10px 20px 10px 300px;
-  padding: 0px 20px;
-  //width: 70%;
-`;
-
-const MessageText = styled.p`
-  font-size: 14px;
-  color: black;
-  margin-top: 10px;
-`;
-
-const MessageLabel = styled.p`
-  font-size: 10px;
-  color: grey;
-  margin-top: -10px;
-`;
-
-function ChatBox({ username, room, Socket }: Props) {
-  const [text, setText] = useState("");
-  const [messages, setMessages] = useState<any[]>([]);
-
-  const dispatchProcess = (encrypt: any, msg: any, cipher: any) => {
-    // process(encrypt, msg, cipher);
-  };
+function ChatBox({ username, room }: Props) {
+  const [text, setText] = useState('');
+  const [messages, setMessages] = useState<ChatMessage[]>([]);
+  const bottomRef = useRef<HTMLDivElement>(null);
 
   useEffect(() => {
-    Socket.on("message", (data: any) => {
-      //decypt the message
-      //const ans = to_Decrypt(data.text, data.username);
-      //dispatchProcess(false, ans, data.text);
-      //console.log(ans);
-      let temp: any[] = messages;
-      temp.push(data);
-      // {
-      //   userId: data.userId,
-      //   username: data.username,
-      //   text: data.text,
-      // });
-      setMessages([...temp]);
-    });
-  }, [Socket]);
+    const handleMessage = (data: ChatMessage) => {
+      setMessages((prev) => [...prev, data]);
+    };
+    socket.on('message', handleMessage);
+    return () => {
+      socket.off('message', handleMessage);
+    };
+  }, []);
+
+  useEffect(() => {
+    bottomRef.current?.scrollIntoView({ behavior: 'smooth' });
+  }, [messages]);
 
   const sendData = () => {
-    if (text !== "") {
-      //encrypt the message here
-      //const ans = to_Encrypt(text);
-      Socket.emit("chat", text);
-      setText("");
+    if (text.trim() !== '') {
+      socket.emit('chat', text.trim());
+      setText('');
     }
   };
-  const messagesEndRef = useRef(null);
-
-  const scrollToBottom = () => {
-    //messagesEndRef.current.scrollIntoView({ behavior: "smooth" });
-  };
-
-  useEffect(scrollToBottom, [messages]);
-
-  //console.log(messages, "mess");
 
   return (
-    <ChatBoxContainer>
-      <div className="user-name">
-        <h2>
-          {username} <span style={{ fontSize: "0.7rem" }}>in {room}</span>
-        </h2>
-      </div>
-      <div className="chat-message">
-        {messages.map((i: any) => {
-          if (i.username === username) {
-            return (
-              <MessageRowSent>
-                <MessageText>{i.text}</MessageText>
-                <MessageLabel>{i.username}</MessageLabel>
-              </MessageRowSent>
-            );
-          } else {
-            return (
-              <MessageRowRecieved>
-                <MessageText>{i.text} </MessageText>
-                <MessageLabel>{i.username}</MessageLabel>
-              </MessageRowRecieved>
-            );
-          }
+    <div className="w-11/12">
+      <h2 className="text-base font-semibold mb-3">
+        {username} <span className="text-xs text-gray-500 font-normal">in {room}</span>
+      </h2>
+
+      <div className="flex flex-col gap-2 max-h-64 overflow-y-auto mb-4 pr-1">
+        {messages.map((msg, i) => {
+          const isMine = msg.username === username;
+          return (
+            <div
+              key={i}
+              className={`flex flex-col border border-gray-300 rounded px-4 py-2 max-w-[70%] ${
+                isMine ? 'self-end items-end' : 'self-start items-start'
+              }`}
+            >
+              <p className="text-sm text-gray-900">{msg.text}</p>
+              <p className="text-xs text-gray-400">{msg.username}</p>
+            </div>
+          );
         })}
-        <div ref={messagesEndRef} />
+        <div ref={bottomRef} />
       </div>
-      <div className="send">
+
+      <div className="flex gap-2">
         <input
-          placeholder="enter your message"
+          className="flex-1 border border-gray-300 rounded px-3 py-2 text-sm focus:outline-none focus:ring-2 focus:ring-blue-400"
+          placeholder="Type a message…"
           value={text}
           onChange={(e) => setText(e.target.value)}
-          onKeyPress={(e) => {
-            if (e.key === "Enter") {
-              sendData();
-            }
+          onKeyDown={(e) => {
+            if (e.key === 'Enter') sendData();
           }}
-        ></input>
-        <button onClick={sendData}>Send</button>
+        />
+        <button
+          className="bg-blue-600 text-white px-4 py-2 rounded text-sm hover:bg-blue-700 transition-colors"
+          onClick={sendData}
+        >
+          Send
+        </button>
       </div>
-    </ChatBoxContainer>
+    </div>
   );
 }
+
 export default ChatBox;
