@@ -1,89 +1,55 @@
-import React, { useState, useEffect } from "react";
-import { advanceGameStage, initalizeGame } from "../Logic/gameplay";
-import Table from "./Table";
-import Player from "./Player";
-import styled from "@emotion/styled";
-import { Poker, gameStages } from "../Logic/types";
+import Table from './Table';
+import Player from './Player';
+import { useGameStore } from '../store/gameStore';
+import { GAME_STAGES } from '@pixelpoker/shared';
+import socket from '../socket';
+import type { GameAction } from '@pixelpoker/shared';
 
-interface Props {
-  Socket: any;
-  username?: string;
-  room?: string;
-}
+function GameContainer() {
+  const game = useGameStore((state) => state.game);
+  const myPlayerIndex = useGameStore((state) => state.myPlayerIndex);
 
-let PlayerContainer = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: space-around;
-  align-items: center;
-  margin: 16px 0px;
-`;
+  if (!game) {
+    return (
+      <div className="flex justify-center items-center h-64 text-gray-500 text-lg">
+        Waiting for game to start…
+      </div>
+    );
+  }
 
-let RootContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  margin: 16px 0px;
-  text-align: center;
-`;
-
-function GameContainer({ Socket, username, room }: Props) {
-  const [game, setGame] = useState(initalizeGame);
-
-  useEffect(() => {
-    Socket.on("updateGame", (data: any) => {
-      setGame(data);
-    });
-  }, [Socket]);
-
-  useEffect(() => {
-    if (
-      (game.currentBet > 0 &&
-        game.currentBet === game.players[game.actionOn].lastBet) ||
-      (game.currentBet === 0 && game.players[game.actionOn].checked)
-    ) {
-      const result = advanceGameStage(game);
-      setGame(result);
-    }
-  }, [game]);
-
-  const nextStep = () => {
-    const result = advanceGameStage(game);
-    setGame(result);
-  };
-
-  const saveNextGameState = (newGame: Poker) => {
-    setGame(newGame);
+  const advanceStage = () => {
+    const action: GameAction = { type: 'advance', playerIndex: -1 };
+    socket.emit('gameAction', action);
   };
 
   return (
-    <RootContainer>
-      <Table
-        tableCards={game.tableCards}
-        pot={game.pot}
-        currentBet={game.currentBet}
-      />
+    <div className="flex flex-col justify-center my-4 text-center">
+      <Table tableCards={game.tableCards} pot={game.pot} currentBet={game.currentBet} />
       <div>
-        <button onClick={nextStep}>
-          {game.stage < 4
-            ? `Deal ${gameStages[game.stage]}`
-            : gameStages[game.stage]}
+        <button
+          className="bg-green-600 text-white px-5 py-2 rounded my-2 hover:bg-green-700 transition-colors"
+          onClick={advanceStage}
+        >
+          {game.stage < 4 ? `Deal ${GAME_STAGES[game.stage]}` : GAME_STAGES[game.stage]}
         </button>
-        <PlayerContainer>
-          {game.players.map((player: any, index: number) => (
+
+        <div className="flex flex-row justify-around items-start my-4 flex-wrap gap-4">
+          {game.players.map((player, index) => (
             <Player
+              key={player.id}
               player={player}
               index={index}
-              game={game}
-              saveGame={saveNextGameState}
               dealer={game.dealer}
               winner={game.winner}
+              actionOn={game.actionOn}
+              currentBet={game.currentBet}
+              numPlayers={game.players.length}
+              isMe={myPlayerIndex === index}
             />
           ))}
-        </PlayerContainer>
-        {/* <button onClick={createPlayer}>Add Player</button> */}
+        </div>
       </div>
-    </RootContainer>
+    </div>
   );
 }
 
