@@ -1,42 +1,26 @@
-const suites = ["c", "h", "s", "d"];
-const cardinality = [2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14];
-const label = [
-  "2",
-  "3",
-  "4",
-  "5",
-  "6",
-  "7",
-  "8",
-  "9",
-  "10",
-  "J",
-  "Q",
-  "K",
-  "A",
-];
+import type { CardType } from './types';
 
-export const generateDeck = () => {
-  const deck = suites.flatMap((suite) =>
-    cardinality.map((value, index) => ({
-      suite,
-      value: label[index] + suite,
-      label: label[index],
-    }))
-  );
+const SUITES = ['c', 'h', 's', 'd'] as const;
+const LABELS = ['2', '3', '4', '5', '6', '7', '8', '9', '10', 'J', 'Q', 'K', 'A'];
 
-  const shuffle1 = shuffle(deck);
-  const shuffle2 = shuffle(deck);
-  const shuffle3 = shuffle(deck);
-
-  return shuffle3;
+const shuffle = (array: CardType[]): CardType[] => {
+  const result = array.slice();
+  for (let i = result.length - 1; i > 0; i--) {
+    const j = Math.floor(Math.random() * (i + 1));
+    [result[i], result[j]] = [result[j], result[i]];
+  }
+  return result;
 };
 
-const shuffle = (array: Array<any>): Array<any> => {
-  const newArray = array.slice(0);
-  for (let i = newArray.length - 1; i > 0; i--) {
-    let j = Math.floor(Math.random() * (i + 1));
-    [newArray[i], newArray[j]] = [newArray[j], newArray[i]];
-  }
-  return newArray;
+export const generateDeck = (): CardType[] => {
+  const deck = SUITES.flatMap((suite) =>
+    LABELS.map((label) => ({
+      suite,
+      // pokersolver expects e.g. "Ac", "2h", "10s"
+      value: label + suite,
+      label,
+    }))
+  );
+  // Three independent shuffles for good randomisation
+  return shuffle(shuffle(shuffle(deck)));
 };
